FreeBSD fixes
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@663 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_boot.c
@@ -232,7 +232,7 @@ mx_get_system_boot_time( struct timespec *system_boot_timespec )
 
 /*---------------------- MacOS X ----------------------*/
 
-#elif defined(OS_MACOSX)
+#elif defined(OS_MACOSX) || defined(__FreeBSD__)
 
 #include <sys/sysctl.h>
 
@@ -272,7 +272,7 @@ mx_get_system_boot_time( struct timespec *system_boot_timespec )
 
 #if MX_BOOT_DEBUG
 	MX_DEBUG(-2,("%s: system_boot_timespec = (%lu,%ld)", fname,
-			system_boot_timespec->tv_sec,
+		(unsigned long) system_boot_timespec->tv_sec,
 			system_boot_timespec->tv_nsec));
 #endif
